+++ smartplate/App.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 
 // Screens
-import PlateTabs from './stack'
+import PlateTabs from './tabs'
 
 // Root
 const App = () => {
